PHPStan error level8 applied

+++ phplib/keyvalue_file.class.php
@@ -3,23 +3,75 @@ namespace KeyValueFile;
 
 class KeyValueFile {
 	
+	/**
+     * @var bool
+     */
 	var $lock = false;
+
+	/**
+     * @var string
+     */
 	var $salt = 'kvfc';
+
+	/**
+     * @var int|bool
+     */
 	var $expires = false;
+
+	/**
+     * @var int
+     */
 	var $expire_span = 0;
+
+	/**
+     * @var int
+     */
 	var $path_depth = 2;
+
+	/**
+     * @var string
+     */
 	var $path = './';
+
+	/**
+     * @var string|null
+     */
 	var $tmp_path = null;
+
+	/**
+     * @var string
+     */
 	var $key = '';
+
+	/**
+     * @var string
+     */
 	var $key_path = './';
+
+	/**
+     * @var bool
+     */
 	var $hashing = true;
+
+	/**
+     * @var string|null
+     */
 	var $path_key_path;
+
+	/**
+     * @var string|null
+     */
 	var $key_fullpath;
+
+	/**
+     * @var bool
+     */
 	var $is_hashed = false;
-//	var $mkdir_permission = 777;
-///	var $tmp_path = '/var/tmp/';
 	
-	
+	/**
+	 * @param string $path
+	 * @param array<string, string|int|bool> $option
+	 */
 	function __construct( $path, $option = null ) {
 		$this->path = $path;
 		// パスの書き込みチェック
@@ -38,10 +90,10 @@ class KeyValueFile {
 			$this->path_depth = 0;
 		}
 		if ( isset($option['umask']) ) {
-			umask($option['umask']);
+			umask((int) $option['umask']);
 		}
-		if ( isset($option['expires']) ) {
-			$this->expires = $option['expires'];
+		if ( isset($option['expires']) && $option['expires'] !== false) {
+			$this->expires = (int) $option['expires'];
 		}
 		
 		// 書き込みファイルのmvロック用ディレクトリ作成
@@ -52,31 +104,50 @@ class KeyValueFile {
 		
 	}
 	
+	/**
+	 * @param int $span_time
+	 * @return void
+	 */
 	function set_expire_span($span_time) {
 		$this->expire_span = $span_time;
 	}
 	
-	
+	/**
+	 * @param string $salt
+	 * @return void
+	 */
 	function set_salt($salt) {
 		$this->salt = $salt;
 	}
 	
+	/**
+	 * @param string $hash_key
+	 * @return bool
+	 */
 	function has_hash_key($hash_key) {
 		$this->is_hashed = true;
 		return $this->has_key($hash_key);
 	}
 	
+	/**
+	 * @param string $key
+	 * @return bool
+	 */
 	function has_key($key) {
 		$this->conv_key_path($key);
 		return file_exists($this->path_key_path . $this->key);
 	}
 	
+	/**
+	 * @param string $key
+	 * @return void
+	 */
 	function conv_key_path($key) {
 		$key = $this->hashing ? sha1($key . $this->salt) : $key;
 		$key_path = '';
 		if ( $this->path_depth > 0 ) {
 			foreach ( range(0, $this->path_depth - 1 )as $i ) {
-				$key_path .= substr($key, $i * 2, 2) . '/';
+				$key_path .= substr($key, (int) $i * 2, 2) . '/';
 			}
 		}
 		
@@ -84,9 +155,13 @@ class KeyValueFile {
 		$this->key_path = $key_path;
 		$this->path_key_path = $this->path . '/' . $key_path;
 		$this->key_fullpath = $this->path . '/' . $key_path . $key;
-//		dde(get_object_vars($this));
 	}
 	
+	/**
+	 * @param string $key
+	 * @param mixed $value
+	 * @return bool
+	 */
 	function set_keyvalue($key, $value) {
 		if ( $this->key == '' ) {
 			$this->conv_key_path($key);
@@ -96,7 +171,7 @@ class KeyValueFile {
 //		if ( ! file_exists($lock_dir) ) {
 //		}
 		
-		if ( ! file_exists($this->path_key_path) ) {
+		if ( $this->path_key_path !== null && !file_exists($this->path_key_path) ) {
 			// パーミッションはconstructorのumaskで調節する
 			mkdir($this->path_key_path, 0777, true);
 		}
@@ -121,11 +196,18 @@ class KeyValueFile {
 		return true;
 	}
 	
+	/**
+	 * @param string $hash_key
+	 * @return mixed
+	 */
 	function get_hash_keyvalue($hash_key) {
 		return $this->get_keyvalue($hash_key);
 	}
 	
-	
+	/**
+	 * @param string $key
+	 * @return mixed|bool
+	 */
 	function get_keyvalue($key) {
 //		$this->set_is_hashed(true);
 		if ( $this->has_key($key) ) {
@@ -137,7 +219,10 @@ class KeyValueFile {
 		}
 	}
 	
-	
+	/**
+	 * @param string $key
+	 * @return bool
+	 */
 	function is_cache_available($key) {
 		$cache = [];
 		if ( $this->has_key($key) ) {
@@ -148,22 +233,40 @@ class KeyValueFile {
 		return time() > $cache['expires'] ? false : true;
 	}
 	
-	
+	/**
+	 * @return mixed
+	 */
 	function get_raw_cache() {
-		return unserialize(file_get_contents($this->key_fullpath));
+		if ($this->key_fullpath) {
+			return unserialize((string) file_get_contents($this->key_fullpath));
+		}
 	}
 	
+	/**
+	 * @return int|false
+	 */
 	function get_cache_mtime() {
-		return filemtime($this->key_fullpath);
+		if ($this->key_fullpath) {
+			return filemtime($this->key_fullpath);
+		} else {
+			return false;
+		}
+		
 	}
 	
+	/**
+	 * @param string $key
+	 * @return bool
+	 */
 	function remove_file($key) {
-		if ( ! $this->has_key($key) ) return false;
-		
+		if ( !$this->has_key($key) || !$this->key_fullpath) return false;
 		// remove dirも入れるか？
 		return unlink($this->key_fullpath);
 	}
 	
+	/**
+	 * @return false
+	 */
 	function _keyError() {
 		echo "Doesn't exist file. $this->key";
 		return false;
